pyproject.toml
[project]
name = "gutenberg"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.22.1",
    "chromadb>=1.5.5",
    "httpx>=0.28.1",
    "matplotlib>=3.10.8",
    "ocrmypdf>=17.4.0",
    "pymupdf>=1.27.2.2",
    "rank-bm25>=0.2.2",
    "runpod>=1.8.2",
    "sentence-transformers>=5.3.0",
    "spacy>=3.8.14",
    "tiktoken>=0.12.0",
    "transformers>=4.40",
    "ebooklib>=0.18",
    "beautifulsoup4>=4.12",
    "python-docx>=1.2.0",
    "sse-starlette>=3.3.4",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
]
